refactor: 경기 진행 함수 파일 분리
